Refactor code by utilizing common functions
