+++ openapi-cli/src/main/java/io/ballerina/openapi/cmd/CmdConstants.java
@@ -206,8 +206,6 @@ public class CmdConstants {
             List.of("2.0", "3.0.0", "3.0.1", "3.0.2", "3.0.3", "3.1.0");
     public static final String DEFAULT_CLIENT_ID = "oas_%s_%s";
     public static final String OPENAPI_ADD_CMD = "add";
-    public static final String OPENAPI_FLATTEN_CMD = "flatten";
-    public static final String OPENAPI_SANITIZE_CMD = "sanitize";
 
     public enum Mode {
         SERVICE,
